requirements.txt
clip==1.0
Flask==2.1.2
numpy==1.22.3
Pillow==9.1.1
Pillow==10.4.0
torch==2.3.1
torchvision==0.18.1
tqdm==4.64.1
piexif==1.1.3
